Refactor recovery node
