+++ nav2_behavior_tree/test/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(plugins/condition)
 add_subdirectory(plugins/decorator)
+add_subdirectory(plugins/control)
